fix: 修复发布后 registry_update 中版本号错误的问题
