+++ tests/github/publish/utils/test_publish_resolve_conflict_pull_requests.py
@@ -319,6 +319,7 @@ async def test_resolve_conflict_pull_requests_plugin(
         type="application",
         supported_adapters=["~onebot.v11"],
     )
+    mock_test_result.version = "1.0.0"
     mock_docker = mocker.patch("src.providers.docker_test.DockerPluginTest.run")
     mock_docker.return_value = mock_test_result
