+++ tests/utils/store_test/test_utils.py
@@ -16,7 +16,7 @@ async def test_load_json_failed(mocked_api: MockRouter):
 
 async def test_get_pypi_data_failed(mocked_api: MockRouter):
     """获取 PyPI 数据失败"""
-    from src.providers.store_test.utils import get_pypi_data
+    from src.providers.utils import get_pypi_data
 
     with pytest.raises(ValueError, match="获取 PyPI 数据失败："):
         get_pypi_data("project_link_failed")
